requirements.txt
attrs
dramatiq[redis]
dynaconf
Flask
flask-bootstrap
flask-wtf
gevent
gunicorn
pika
psycopg2
pytest
redis
requests
sqlalchemy
tenacity
